Fix epoll_wait and epoll_pwait decoding
* desc.c (epoll_wait_common): Print "maxevents" and "timeout" arguments
as integers.

Reported-by: Марк Коренберг <socketpair@gmail.com>

+++ desc.c
@@ -777,7 +777,7 @@ epoll_wait_common(struct tcb *tcp)
 			tprints("{...}");
 #endif
 		}
-		tprintf(", %ld, %ld", tcp->u_arg[2], tcp->u_arg[3]);
+		tprintf(", %d, %d", (int) tcp->u_arg[2], (int) tcp->u_arg[3]);
 	}
 }
